requirements.txt
accelerate==1.10.1
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.71.0
anyio==4.10.0
apache-tvm-ffi==0.1.3
astor==0.8.1
attrs==25.3.0
beautifulsoup4==4.13.5
black==25.1.0
blake3==1.0.5
Brotli==1.1.0
cachetools==6.2.0
cbor2==5.7.0
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
compressed-tensors==0.12.2
cssselect==1.3.0
cuda-bindings==13.0.3
cuda-pathfinder==1.3.2
cuda-python==13.0.3
cupy-cuda12x==13.6.0
dataclasses-json==0.6.7
depyf==0.20.0
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
einops==0.8.1
email-validator==2.3.0
fastapi==0.121.3
fastapi-cli==0.0.10
fastapi-cloud-cli==0.1.5
fastrlock==0.8.3
filelock==3.19.1
flash_attn==2.8.3
flashinfer-python==0.5.2
frozenlist==1.7.0
fsspec==2024.6.1
gguf==0.17.1
google_search_results==2.4.2
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
identify==2.6.13
idna==3.10
interegular==0.3.3
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-text-splitters==0.3.11
langsmith==0.4.21
lark==1.2.2
llguidance==1.3.0
llvmlite==0.44.0
lm-format-enforcer==0.11.3
loguru==0.7.3
lxml==6.0.1
lxml_html_clean==0.4.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mistral_common==1.8.5
model-hosting-container-standards==0.1.9
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.4
mypy_extensions==1.1.0
networkx==3.3
ninja==1.13.0
nodeenv==1.9.1
numba==0.61.2
numpy==2.1.2
nvidia-cublas==13.0.0.19
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti==13.0.48
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc==13.0.48
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime==13.0.48
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-cu13==9.13.0.50
nvidia-cudnn-frontend==1.14.1
nvidia-cufft==12.0.0.15
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile==1.15.0.42
nvidia-cufile-cu12==1.13.1.3
nvidia-curand==10.4.0.35
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver==12.0.3.29
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse==12.6.2.49
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-cusparselt-cu13==0.8.0
nvidia-cutlass-dsl==4.3.0
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.5
nvidia-nccl-cu13==2.27.7
nvidia-nvjitlink==13.0.39
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvshmem-cu13==3.3.24
nvidia-nvtx==13.0.39
nvidia-nvtx-cu12==12.8.90
openai==1.104.2
openai-harmony==0.0.4
opencv-python-headless==4.12.0.88
orjson==3.11.3
outlines_core==0.2.11
packaging==25.0
partial-json-parser==0.2.1.1.post6
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.4.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
py-cpuinfo==9.0.0
pybase64==1.4.2
pycares==4.11.0
pycountry==24.6.1
pycparser==2.22
pydantic==2.12.4
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
pyzmq==27.0.2
ray==2.49.1
readability-lxml==0.8.4.1
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
rpds-py==0.27.1
safetensors==0.6.2
scipy==1.16.1
sentencepiece==0.2.1
sentry-sdk==2.35.2
setproctitle==1.3.6
shellingham==1.5.4
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
soxr==0.5.0.post1
SQLAlchemy==2.0.43
starlette==0.50.0
supervisor==4.3.0
sympy==1.13.3
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.22.0
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.56.0
triton==3.5.0
typer==0.17.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uv==0.8.15
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.34.0
vllm==0.11.2
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
xformers==0.0.33.post1
xgrammar==0.1.25
yarl==1.20.1
zstandard==0.24.0
